# packages/api-schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nature Explorer API",
    "version": "0.1.0",
    "description": "Generated at startup from the TypeBox schemas of the registered routes. Every non-2xx response uses the shared `Error` schema and echoes the request id in `x-request-id`. `/v1/me*` and `/v1/auth/logout` need a bearer access token (`bearerAuth`)."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from POST /v1/auth/apple or /v1/auth/refresh (HS256, 15 minutes)"
      }
    },
    "schemas": {
      "Error": {
        "additionalProperties": false,
        "description": "Shared error envelope returned by every non-2xx response. `code` is a stable machine-readable identifier in SCREAMING_SNAKE_CASE; `message` is for developers, never shown verbatim to players; `details` carries validation issues or other structured context.",
        "type": "object",
        "required": ["error", "requestId"],
        "properties": {
          "error": {
            "additionalProperties": false,
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "pattern": "^[A-Z][A-Z0-9_]+$",
                "type": "string",
                "example": "VALIDATION_FAILED"
              },
              "message": { "type": "string" },
              "details": { "additionalProperties": true, "type": "object", "properties": {} }
            }
          },
          "requestId": {
            "description": "Fastify request id, also returned in the `x-request-id` response header",
            "type": "string"
          }
        }
      },
...
      "Me": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "displayName",
          "factionId",
          "factionChangedAt",
          "factionChangeAvailableAt",
          "xp",
          "level",
          "role",
          "createdAt",
          "suggestedFactionId"
        ],
        "properties": {
          "id": { "format": "uuid", "type": "string" },
          "displayName": { "type": "string" },
          "factionId": { "type": ["null", "integer"] },
          "factionChangedAt": { "format": "date-time", "type": ["null", "string"] },
          "factionChangeAvailableAt": {
            "format": "date-time",
            "description": "When the next faction change is allowed; null when a change is allowed now (or no faction is set yet)",
            "type": ["null", "string"]
          },
          "xp": { "minimum": 0, "type": "integer" },
          "level": { "minimum": 1, "type": "integer" },
          "role": { "type": "string", "enum": ["player", "tester", "admin"] },
          "createdAt": { "format": "date-time", "type": "string" },
          "suggestedFactionId": {
            "description": "Faction with the fewest active members (tie → lowest id)",
            "type": "integer"
          }
        }
      },
...
      "AuthAppleRequest": {
        "additionalProperties": false,
        "type": "object",
        "required": ["identityToken"],
        "properties": {
          "identityToken": {
            "minLength": 1,
            "maxLength": 8192,
            "description": "Apple identity token (JWT) from ASAuthorizationAppleIDCredential.identityToken",
            "type": "string"
          },
          "authorizationCode": {
            "minLength": 1,
            "maxLength": 2048,
            "description": "Apple authorization code; accepted and ignored in feature 002 (used by the token-revocation flow of feature 009)",
            "type": "string"
          },
          "fullName": {
            "additionalProperties": false,
            "description": "Present only on the first authorization, when Apple hands the name to the app",
            "type": "object",
            "properties": {
              "givenName": { "maxLength": 100, "type": "string" },
              "familyName": { "maxLength": 100, "type": "string" }
            }
          }
        }
      },
      "TokenPair": {
        "additionalProperties": false,
        "type": "object",
        "required": ["accessToken", "accessExpiresAt", "refreshToken", "refreshExpiresAt"],
        "properties": {
          "accessToken": { "type": "string" },
          "accessExpiresAt": { "format": "date-time", "type": "string" },
          "refreshToken": { "type": "string" },
          "refreshExpiresAt": { "format": "date-time", "type": "string" }
        }
      },
      "AuthResponse": {
        "additionalProperties": false,
        "type": "object",
        "required": ["tokens", "me", "isNewUser", "restored"],
        "properties": {
          "tokens": { "$ref": "#/components/schemas/TokenPair" },
          "me": { "$ref": "#/components/schemas/Me" },
          "isNewUser": { "type": "boolean" },
          "restored": {
            "description": "True when an account marked deleted was reactivated by this sign-in",
            "type": "boolean"
          }
        }
      },
      "RefreshRequest": {
        "additionalProperties": false,
        "type": "object",
        "required": ["refreshToken"],
        "properties": { "refreshToken": { "minLength": 1, "maxLength": 256, "type": "string" } }
      },
      "LogoutRequest": {
        "additionalProperties": false,
        "type": "object",
        "required": ["refreshToken"],
        "properties": { "refreshToken": { "minLength": 1, "maxLength": 256, "type": "string" } }
      }
    }
  },
  "paths": {
...
    "/v1/auth/apple": {
      "post": {
        "operationId": "signInWithApple",
        "summary": "Sign in (or up) with an Apple identity token",
        "tags": ["auth"],
        "description": "Verifies `identityToken` against Apple's JWKS (issuer `https://appleid.apple.com`, audience = the app's bundle identifier(s)), then creates or resumes the account keyed by Apple's `sub`. E-mail and `fullName` are stored only when the account is created; later calls ignore them. An account marked deleted within the last 30 days is restored (`restored: true`). Rate limited: 20 requests per minute per client address.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AuthAppleRequest" } }
          }
        },
        "security": [],
        "responses": {
          "200": {
            "description": "Tokens for the (new or existing) account and the player's profile",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } }
            }
          },
          "400": {
            "description": "Request validation failed (code VALIDATION_FAILED)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": {
            "description": "The identity token is invalid (code INVALID_APPLE_TOKEN, `details.reason` in expired | audience | issuer | signature | malformed)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "429": {
            "description": "Too many requests from this client address (code RATE_LIMITED, `details.retryAfterS`); the `retry-after` header is set",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "503": {
            "description": "Apple's key service is unreachable (code APPLE_UNAVAILABLE); retry later",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "refreshTokens",
        "summary": "Rotate a refresh token",
        "tags": ["auth"],
        "description": "Exchanges a valid refresh token for a new token pair and revokes the presented token. Presenting a token that was already rotated revokes every refresh token of the account (code REFRESH_REUSED). Rate limited: 20 requests per minute per client address.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RefreshRequest" } }
          }
        },
        "security": [],
        "responses": {
          "200": {
            "description": "New token pair",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/TokenPair" } }
            }
          },
          "400": {
            "description": "Request validation failed (code VALIDATION_FAILED)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": {
            "description": "Unknown, expired or revoked token (INVALID_REFRESH_TOKEN), reuse of a rotated token (REFRESH_REUSED), or deleted account (ACCOUNT_DELETED); the client must sign in again",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "429": {
            "description": "Too many requests from this client address (code RATE_LIMITED, `details.retryAfterS`); the `retry-after` header is set",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "summary": "Revoke the presented refresh token (sign out this device)",
        "tags": ["auth"],
        "description": "Revokes the refresh token if it belongs to the caller. Always succeeds for a valid session, even when the token is already revoked or unknown. Other devices are unaffected.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LogoutRequest" } }
          }
        },
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Signed out" },
          "400": {
            "description": "Request validation failed (code VALIDATION_FAILED)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": {
            "description": "Missing or invalid access token (UNAUTHORIZED), expired token (TOKEN_EXPIRED) or deleted account (ACCOUNT_DELETED)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "429": {
            "description": "Too many requests from this client address (code RATE_LIMITED, `details.retryAfterS`); the `retry-after` header is set",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
...
  },
  "servers": [{ "url": "http://localhost:3000", "description": "Local development (make dev)" }],
...
}